requirements.txt
requests
python-dotenv
mutagen
yt-dlp
